Moved state to correct component
Moved from Addindividual.js to programlist.js where I need the state to be to populate the list.
