+++ src/components/pages/Programs/ImportPrograms/AddIndividual.js
@@ -1,8 +1,5 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { Form, Input, Button } from 'antd';
-import axios from 'axios';
-import { baseUrl } from '../../../../api/index';
-// const baseUrl = "http://localhost:2019";
 
 const layout = {
   wrapperCol: {
@@ -18,22 +15,6 @@ const tailLayout = {
 
 export const AddIndividual = props => {
   const { inputData, setInputData } = props;
-  const [activities, setActivities] = useState([]);
-
-  const clubid = 20;
-  // this will be handled in context state, hard coded for testing now.
-
-  useEffect(() => {
-    let tokenObj = JSON.parse(localStorage.getItem('okta-token-storage'));
-    axios
-      .get(`${baseUrl}/clubs/clubs`, {
-        headers: {
-          Authorization: `Bearer ${tokenObj.accessToken.accessToken}`,
-        },
-      })
-      .then(res => setActivities(res.data))
-      .catch(e => console.log(e));
-  }, []);
 
   const onFinish = values => {
     console.warn(values);
